Rename FileToWasbOperator to LocalFilesystemToWasbOperator (#18109)
