+++ airflow/providers/microsoft/azure/example_dags/example_local_to_wasb.py
@@ -19,13 +19,13 @@ import os
 
 from airflow.models import DAG
 from airflow.providers.microsoft.azure.operators.wasb_delete_blob import WasbDeleteBlobOperator
-from airflow.providers.microsoft.azure.transfers.file_to_wasb import FileToWasbOperator
+from airflow.providers.microsoft.azure.transfers.local_to_wasb import LocalFilesystemToWasbOperator
 from airflow.utils.dates import days_ago
 
 PATH_TO_UPLOAD_FILE = os.environ.get('AZURE_PATH_TO_UPLOAD_FILE', 'example-text.txt')
 
-with DAG("example_file_to_wasb", schedule_interval="@once", start_date=days_ago(2)) as dag:
-    upload = FileToWasbOperator(
+with DAG("example_local_to_wasb", schedule_interval="@once", start_date=days_ago(2)) as dag:
+    upload = LocalFilesystemToWasbOperator(
         task_id="upload_file", file_path=PATH_TO_UPLOAD_FILE, container_name="mycontainer", blob_name='myblob'
     )
     delete = WasbDeleteBlobOperator(task_id="delete_file", container_name="mycontainer", blob_name="myblob")
