+++ dev/provider_packages/prepare_provider_packages.py
@@ -2117,6 +2117,7 @@ KNOWN_DEPRECATED_MESSAGES: Set[Tuple[str, str]] = {
 # we imported it directly during module walk by the importlib library
 KNOWN_DEPRECATED_DIRECT_IMPORTS: Set[str] = {
     "This module is deprecated. Please use `airflow.providers.amazon.aws.hooks.dynamodb`.",
+    "This module is deprecated. Please use `airflow.providers.microsoft.azure.transfers.local_to_wasb`.",
     "This module is deprecated. Please use `airflow.providers.tableau.operators.tableau_refresh_workbook`.",
     "This module is deprecated. Please use `airflow.providers.tableau.sensors.tableau_job_status`.",
     "This module is deprecated. Please use `airflow.providers.tableau.hooks.tableau`.",
